refactor(frontend): extract GoalFilters and GoalList from goals page (#58 #210)
- Extract the editor filter/search header into `GoalFilters.svelte`.
- Extract the editor goal grid into `GoalList.svelte`.
- Update `routes/goals/+page.svelte` to use the new components.
- Remove duplicated `filteredGoals` logic and editor styles.

Closes #58, closes #210.
